added possibility to not to show SelectSource screen:
SelectSource screen is opened when not-resolved video
is resolved and it has more than one video resolved.

It can be usefull to not to open this screen when it
takes long time to resolve not-resolved video (captcha, timeout).

So you don't loose direct video link. For example when you
play video item, next you jump out of player, you don't need to
resolve video item again, but use already resolved direct video link
for playback or downloading