player: fix possible GS in AudioSelection
AudioSelection can be used also for subtitles selection. When
in AudioSelection are detected subtitles, it shows  subtitles menu so
you can select subtitles. Problem is that AudioSelection also expects
that InfoBarSubtitlesSupport is part of infobar, since it want's to
access it's variables when building subtitles list.
We are not using InfoBarSubtitlesSupport, since we use SubsSupport
instead.

To fix this use custom InfoBarAudioSelectionNoSubtitles which opens
subclassed AudioSelectionNoSubtitles which as name suggest doesn't
return any subtitles, so subtitles menu is never accessible.